Migrating Absolute Time Range fields to their own components
